ONB-71 refactor: mail 전송 유틸 함수로 리팩토링
